\name{Ds}
\alias{Ds}
\docType{data}
\title{
A list of ecological distance matrices for the gut microbiome data (Yanai et al., 2024, Nat Commun)
}
\description{
A list of ecological distance matrices for the gut microbiome data (Yanai et al., 2024, Nat Commun)
}
\usage{data("Ds")}
\format{
  The format is:
List of 7
 $ : num [1:48, 1:48] 0 0.568 0.574 0.456 0.712 ...
  ..- attr(*, "dimnames")=List of 2
  .. ..$ : chr [1:48] "JMat-baseline-DCGG" "JMat-C3D2-DCGG" "JMat-C3D4-DCGG" "JMat-C3D8-DCGG" ...
  .. ..$ : chr [1:48] "JMat-baseline-DCGG" "JMat-C3D2-DCGG" "JMat-C3D4-DCGG" "JMat-C3D8-DCGG" ...
 $ : num [1:48, 1:48] 0 0.483 0.694 0.506 0.83 ...
  ..- attr(*, "dimnames")=List of 2
  .. ..$ : chr [1:48] "JMat-baseline-DCGG" "JMat-C3D2-DCGG" "JMat-C3D4-DCGG" "JMat-C3D8-DCGG" ...
  .. ..$ : chr [1:48] "JMat-baseline-DCGG" "JMat-C3D2-DCGG" "JMat-C3D4-DCGG" "JMat-C3D8-DCGG" ...
 $ : num [1:48, 1:48] 0 0.338 0.399 0.281 0.489 ...
  ..- attr(*, "dimnames")=List of 2
  .. ..$ : chr [1:48] "S1" "S2" "S3" "S4" ...
  .. ..$ : chr [1:48] "S1" "S2" "S3" "S4" ...
 $ : num [1:48, 1:48] 0 0.452 0.623 0.457 0.702 ...
  ..- attr(*, "dimnames")=List of 2
  .. ..$ : chr [1:48] "S1" "S2" "S3" "S4" ...
  .. ..$ : chr [1:48] "S1" "S2" "S3" "S4" ...
 $ : num [1:48, 1:48] 0 0.351 0.528 0.374 0.606 ...
  ..- attr(*, "dimnames")=List of 2
  .. ..$ : chr [1:48] "S1" "S2" "S3" "S4" ...
  .. ..$ : chr [1:48] "S1" "S2" "S3" "S4" ...
 $ : num [1:48, 1:48] 0 0.266 0.439 0.309 0.505 ...
  ..- attr(*, "dimnames")=List of 2
  .. ..$ : chr [1:48] "S1" "S2" "S3" "S4" ...
  .. ..$ : chr [1:48] "S1" "S2" "S3" "S4" ...
 $ : num [1:48, 1:48] 0 0.203 0.362 0.259 0.413 ...
  ..- attr(*, "dimnames")=List of 2
  .. ..$ : chr [1:48] "S1" "S2" "S3" "S4" ...
  .. ..$ : chr [1:48] "S1" "S2" "S3" "S4" ...
}

\references{
Yanai et al. Short-term periodic restricted feeding elicits metabolome-microbiome signatures with sex dimorphic persistence in primate intervention. Nat Commun, 15(1088), 2024.
}
\examples{
data(Ds)
}
\keyword{datasets}
