\name{KernRBD-package}
\alias{KernRBD-package}
\alias{KernRBD}
\docType{package}
\title{
\packageTitle{KernRBD}
}
\description{
\packageDescription{KernRBD}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{KernRBD}
\packageIndices{KernRBD}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{KernRBD}

Maintainer: \packageMaintainer{KernRBD}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS ~~
~~ in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
