\name{KernRBD-package}
\alias{KernRBD-package}
\alias{KernRBD}
\docType{package}
\title{
\packageTitle{KernRBD}
}
\description{
\packageDescription{KernRBD}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{KernRBD}
\packageIndices{KernRBD}
}
\author{
\packageAuthor{KernRBD}

Maintainer: \packageMaintainer{KernRBD}
}
\references{
Koh H. A general kernel machine comparative analysis framework for randomized block designs. (In Review)
}
